requirements.txt
pandas
nltk
keras
sklearn
pymongo
python-arango
tensorflow